Tools & examples for using gcc/ld linkersets
This change adds an implementation for basic linkersets (linkerset.h),
an implementation for a linkerset that allows proper module
initialization (initializing only the modules that are linked into the
program), and one example for each.

This repository holds examples of using linkersets (sets of unordered
data that can be produced by gcc / ld for arbitrary data.  The initial
examples that are included:

o Module initialization

  Ensure that modules are initialized once, and in the proper order
  based on the DAG produced by modules importing other modules.

  Execute 'make' in the directory, then run:

  ./example

o Simple

  The simple example shoulds that data can be collected, and the
  values present in the final executable are dependent upon what is
  linked into the program.  There are three configurations of the
  program:

    make example circle point

    ./circle
    ./point
    ./example

  The 'circle' and 'point' configurations only contain information
  about the corresponding shape in the linker set, but 'example'
  contains both -- controlled all at compile & link time.